requirements.txt
# He-LMAS: Heterogeneous Latent Manifold Alignment System

# Core ML
torch>=2.0.0
transformers>=4.40.0
accelerate>=0.27.0
bitsandbytes>=0.42.0
datasets>=2.18.0

# Training
pyyaml>=6.0
tqdm>=4.65.0

# HuggingFace Hub
huggingface_hub>=0.21.0

# Logging
wandb>=0.16.0
tensorboard>=2.15.0

# Testing
pytest>=7.4.0
numpy>=1.24.0
matplotlib>=3.8.0
